Fixed bugs

+++ database/migrations/2025_10_13_092550_create_roles_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('roles', function (Blueprint $table) {
             $table->id();
             $table->string('name')->unique();
-            $table->integer('permissionID')->nullable()->default(null);
+            $table->unsignedBigInteger('permissionID')->nullable()->default(null);
             $table->timestamps();
         });
     }
